Move functions into dashboard and change to arrow functions
